Phase 2: Extract custom hooks for Topbar
Extract business logic into reusable custom hooks following Single Responsibility Principle:

Created hooks.ts with two custom hooks:
- useSearchState: Manages search query state and synchronization with Redux
  - Handles complex sync between local state (for input responsiveness) and Redux state
  - Provides handlers for search change, clear, and submit
  - Eliminates manual synchronization logic from component

- useMobileToolbar: Manages mobile toolbar open/close state
  - Encapsulates toggle logic and search clearing behavior
  - Simplifies mobile-specific state management

Updated index.tsx to use new hooks:
- Replaced inline state management with hook calls
- Removed duplicate event handler logic
- Cleaned up imports (removed unused action creators)
- Component now focuses on composition rather than logic

Benefits:
- Better separation of concerns (logic vs presentation)
- Each hook testable in isolation
- Reduced component complexity (removed 40+ lines of logic)
- Improved code reusability
- Maintained SSR-safe conditional render of AltSCycler (typeof window check)

No functional changes - all existing behavior preserved.

Related to CORE-1456
